Fixes to the warnings

+++ main.cpp
@@ -76,24 +76,24 @@ int main()
             auto temp_approx1 = log(PS(k_m1));
             auto temp_true2 = log(ms.PPS(k_m2));
             auto temp_approx2 = log(PS(k_m2));
-            k_pair.erase(k_pair.begin() + n);
+            k_pair.erase(k_pair.begin() + static_cast<int>(n));
             if(abs(temp_true1 - temp_approx1) / temp_true1 > lim and abs(temp_true2 - temp_approx2) / temp_true2 > lim)
             {
-                k_pair.insert(k_pair.begin() + n, std::make_pair(k0, k_m1));
-                k_pair.insert(k_pair.begin() + n + 1, std::make_pair(k_m1, k_m2));
-                k_pair.insert(k_pair.begin() + n + 2, std::make_pair(k_m2, k1));
+                k_pair.insert(k_pair.begin() + static_cast<int>(n), std::make_pair(k0, k_m1));
+                k_pair.insert(k_pair.begin() + static_cast<int>(n) + 1, std::make_pair(k_m1, k_m2));
+                k_pair.insert(k_pair.begin() + static_cast<int>(n) + 2, std::make_pair(k_m2, k1));
                 n += 2;
             }
             else if(abs(temp_true1 - temp_approx1) / temp_true1 > lim and abs(temp_true2 - temp_approx2) / temp_true2 < lim)
             {
-                k_pair.insert(k_pair.begin() + n, std::make_pair(k0, k_m1));
-                k_pair.insert(k_pair.begin() + n + 1, std::make_pair(k_m1, k1));
+                k_pair.insert(k_pair.begin() + static_cast<int>(n), std::make_pair(k0, k_m1));
+                k_pair.insert(k_pair.begin() + static_cast<int>(n) + 1, std::make_pair(k_m1, k1));
                 n += 1;
             }
             else if(abs(temp_true1 - temp_approx1) / temp_true1 < lim and abs(temp_true2 - temp_approx2) / temp_true2 > lim)
             {
-                k_pair.insert(k_pair.begin() + n, std::make_pair(k0, k_m2));
-                k_pair.insert(k_pair.begin() + n + 1, std::make_pair(k_m2, k1));
+                k_pair.insert(k_pair.begin() + static_cast<int>(n), std::make_pair(k0, k_m2));
+                k_pair.insert(k_pair.begin() + static_cast<int>(n) + 1, std::make_pair(k_m2, k1));
                 n += 1;
             }
             PS.insert(k_m1, exp(temp_true1));
